+++ src/components/icons.tsx
@@ -170,6 +170,15 @@ export const ReportingIcon = (props: SVGProps<SVGSVGElement>) => (
   </IconWrapper>
 );
 
+export const NotificationsIcon = (props: SVGProps<SVGSVGElement>) => (
+  <IconWrapper {...props}>
+    <g fill="url(#glass-gradient)">
+       <path d="M18 8A6 6 0 0 0 6 8c0 7-3 9-3 9h18s-3-2-3-9" />
+       <path d="M13.73 21a2 2 0 0 1-3.46 0" />
+    </g>
+  </IconWrapper>
+);
+
 
 export const WhatsNewIcon = (props: SVGProps<SVGSVGElement>) => (
   <IconWrapper {...props}>